requirements.txt
requests==2.18.4
python-gantt==0.6.0
PyYAML==3.12
